# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop-api Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Enter Mobile Number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7777777777"
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "Device Id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "push_notification_id",
                        "in": "query",
                        "description": "push notification id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "device_type",
                        "in": "query",
                        "description": "Android|IOS",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Android"
                    },
                    {
                        "name": "device_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "One Plus Nord"
                    },
                    {
                        "name": "carrier_name",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NA"
                    },
                    {
                        "name": "os_version",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "9"
                    },
                    {
                        "name": "app_version",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1.0"
                    },
                    {
                        "name": "device_country",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NA"
                    },
                    {
                        "name": "time_zone",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Asia/Kolkata"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
        "/api/v1/user/category/list": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Category List",
                "description": "CategoryList",
                "operationId": "CategoryList",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "no of record you already get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "no of record you want to get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column(title)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/item/list": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Item List",
                "description": "ItemList",
                "operationId": "itemList",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "no of record you already get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "no of record you want to get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category Id",
                        "required": false,
                        "schema": {
                            "type": "number"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column(title)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/profile/update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Profile Update",
                "description": "ProfileUpdate",
                "operationId": "ProfileUpdate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a bearer token to access these endpoints",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}